+++ flexmeasures/api/v3_0/tests/test_sensor_schedules_fresh_db.py
@@ -253,7 +253,6 @@ def test_price_sensor_priority(
         f"{other_sensors[sensor_type]}-price-sensor",
     )
 
-    price_sensor_id = None
     sensor_attribute = f"{sensor_type}_price_sensor_id"
     # preparation: ensure the asset actually has the price sensor set as attribute
     if asset_sensor:
@@ -267,7 +266,7 @@ def test_price_sensor_priority(
         setattr(building_asset, sensor_attribute, price_sensor_id)
         fresh_db.session.add(building_asset)
 
-    # Adding unused sensor to context (e.g consumption price sensor if we test production sensor)
+    # Adding unused sensor to context (e.g. consumption price sensor if we test production sensor)
     message["flex-context"] = {
         unused_sensor: add_market_prices_fresh_db["epex_da"].id,
         "site-power-capacity": "1 TW",  # should be big enough to avoid any infeasibilities
@@ -391,6 +390,151 @@ def test_inflexible_device_sensors_priority(
         # Counting how many times power values (for inflexible sensors) were fetched (gives us the number of sensors)
         call_args = mock_storage_get_power_values.call_args_list
         assert len(call_args) == expect_sensor_num
+
+
+@pytest.mark.parametrize(
+    "include_consumption_breach",
+    [
+        False,
+        True,
+    ],
+)
+@pytest.mark.parametrize(
+    "requesting_user", ["test_prosumer_user@seita.nl"], indirect=True
+)
+def test_multiple_contracts(
+    app,
+    fresh_db,
+    add_market_prices_fresh_db,
+    add_battery_assets_fresh_db,
+    battery_soc_sensor_fresh_db,
+    include_consumption_breach,
+    requesting_user,
+):
+    """Check planning against an energy contract, a breach contract and a peak contract."""
+    message, asset_name = (
+        message_for_trigger_schedule(
+            with_targets=True, use_time_window=True, use_perfect_efficiencies=True
+        ),
+        "Test battery",
+    )
+    message["force_new_job_creation"] = True
+
+    if include_consumption_breach:
+        # Given the soc-model defined in message_for_trigger_schedule,
+        # we'll need to breach this site_consumption_capacity to reach the target:
+        # target = 25 kWh
+        # start = 12.1 kWh
+        # delta = 12.9 kWh
+        # time to reach target = 46.75 hours ( from 2015-01-01T00:00:00+01:00 to 2015-01-02T22:45:00+01:00)
+        # minimum power required = 12.9 / 46.75 ≃ 0.276 kW (assuming 100% efficiencies)
+        site_consumption_capacity = 0.25
+    else:
+        # we won't need to breach this site_consumption_capacity to reach the target
+        site_consumption_capacity = 1
+
+    price_sensor_id = add_market_prices_fresh_db["epex_da"].id
+    message["flex-context"] = {
+        "consumption-price": {"sensor": price_sensor_id},
+        "production-price": {"sensor": price_sensor_id},
+        "site-power-capacity": "2 MW",  # should be big enough to avoid any infeasibilities
+        "site-consumption-capacity": f"{site_consumption_capacity} kW",
+        "site-consumption-breach-price": "1000 EUR/kW",
+        "site-production-breach-price": "1000 EUR/kW",
+        "site-peak-consumption": "20 kW",
+        "site-peak-production": "20 kW",
+        "site-peak-consumption-price": "260 EUR/MW",
+        "site-peak-production-price": "260 EUR/MW",
+    }
+
+    sensor = (
+        Sensor.query.filter(Sensor.name == "power")
+        .join(GenericAsset, GenericAsset.id == Sensor.generic_asset_id)
+        .filter(GenericAsset.name == asset_name)
+        .one_or_none()
+    )
+
+    with app.test_client() as client:
+        trigger_schedule_response = client.post(
+            url_for("SensorAPI:trigger_schedule", id=sensor.id),
+            json=message,
+        )
+        print("Server responded with:\n%s" % trigger_schedule_response.json)
+        assert trigger_schedule_response.status_code == 200
+        job_id = trigger_schedule_response.json["schedule"]
+
+    # process the scheduling queue
+    work_on_rq(
+        app.queues["scheduling"], exc_handler=handle_scheduling_exception, max_jobs=1
+    )
+    job = Job.fetch(job_id, connection=app.queues["scheduling"].connection)
+    assert job.is_finished is True
+
+    # First, make sure the scheduler data source is now there
+    job.refresh()  # catch meta info that was added on this very instance
+    scheduler_source = get_data_source_for_job(job)
+    assert scheduler_source is not None
+
+    # try to retrieve the schedule through the /sensors/<id>/schedules/<job_id> [GET] api endpoint
+    get_schedule_response = client.get(
+        url_for("SensorAPI:get_schedule", id=sensor.id, uuid=job_id),
+        query_string={"duration": "PT48H"},
+    )
+    print("Server responded with:\n%s" % get_schedule_response.json)
+    assert get_schedule_response.status_code == 200
+    power_values = get_schedule_response.json["values"]
+    start = get_schedule_response.json["start"]
+    duration = get_schedule_response.json["duration"]
+
+    consumption_schedule = pd.Series(
+        data=power_values,
+        index=pd.date_range(
+            start=start,
+            end=pd.Timestamp(start) + pd.Timedelta(duration),
+            freq="15min",
+            inclusive="left",
+        ),
+    )
+
+    start_soc = message["flex-model"]["soc-at-start"] / 1000  # in MWh
+    roundtrip_efficiency = (
+        float(message["flex-model"]["roundtrip-efficiency"].replace("%", "")) / 100.0
+    )
+    storage_efficiency = (
+        float(message["flex-model"]["storage-efficiency"].replace("%", "")) / 100.0
+    )
+    soc_targets = message["flex-model"].get("soc-targets")
+
+    soc_schedule = integrate_time_series(
+        consumption_schedule,
+        start_soc,
+        up_efficiency=roundtrip_efficiency**0.5,
+        down_efficiency=roundtrip_efficiency**0.5,
+        storage_efficiency=storage_efficiency,
+        decimal_precision=6,
+    )
+    print(consumption_schedule)
+    print(soc_schedule)
+
+    # Check for consumption breaches
+    if include_consumption_breach:
+        # The minimum power level to reach the target is 0.276 kW; higher breaches cost more and thus are avoided
+        assert all(v <= 0.000276 for v in consumption_schedule)
+        # Check for consumption breaches over 0.25 kW
+        assert any(v > site_consumption_capacity / 1000 for v in consumption_schedule)
+    else:
+        # Check for absence of consumption breaches over 1 kW, i.e. any breach costs is avoided
+        assert all(v <= site_consumption_capacity / 1000 for v in consumption_schedule)
+
+    # Check for absence of extra production peaks over 20 kW, i.e. any peak costs are avoided
+    assert all(v >= -0.02 for v in consumption_schedule)
+
+    # Check target is met
+    for target in soc_targets:
+        assert (
+            int(soc_schedule[target.get("datetime", target.get("end"))] * 1000)
+            == target["value"]
+        )
 
 
 def setup_inflexible_device_sensors(fresh_db, asset, sensor_name, sensor_num):