+++ flexmeasures/data/models/planning/utils.py
@@ -18,6 +18,7 @@ from flexmeasures.data.models.planning.exceptions import (
 from flexmeasures import Asset
 from flexmeasures.data.queries.utils import simplify_index
 
+from flexmeasures.utils.flexmeasures_inflection import capitalize, pluralize
 from flexmeasures.utils.unit_utils import ur, convert_units
 from pint.errors import UndefinedUnitError, DimensionalityError
 
@@ -135,41 +136,75 @@ def get_prices(
         raise UnknownPricesException(
             f"Prices unknown for planning window. (sensor {price_sensor.id})"
         )
-    elif (
-        nan_prices.any()
-        or pd.Timestamp(price_df.index[0]).tz_convert("UTC")
+    else:
+        price_df = extend_to_edges(
+            df=price_df,
+            query_window=query_window,
+            resolution=resolution,
+            sensor=price_sensor,
+            allow_trimmed_query_window=allow_trimmed_query_window,
+        )
+    return price_df, query_window
+
+
+def extend_to_edges(
+    df: pd.DataFrame | pd.Series,
+    query_window: tuple[datetime, datetime],
+    resolution: timedelta,
+    kind_of_values: str = "price",
+    sensor: Sensor = None,
+    allow_trimmed_query_window: bool = False,
+):
+    """Values are extended to the edges of the query window.
+
+    - The first available value serves as a naive backcasts.
+    - The last available value serves as a naive forecast.
+    """
+
+    nan_values = df.isnull().values
+    if (
+        nan_values.any()
+        or pd.Timestamp(df.index[0]).tz_convert("UTC")
         != pd.Timestamp(query_window[0]).tz_convert("UTC")
-        or pd.Timestamp(price_df.index[-1]).tz_convert("UTC") + resolution
+        or pd.Timestamp(df.index[-1]).tz_convert("UTC") + resolution
         != pd.Timestamp(query_window[-1]).tz_convert("UTC")
     ):
         if allow_trimmed_query_window:
-            first_event_start = price_df.first_valid_index()
-            last_event_end = price_df.last_valid_index() + resolution
+            first_event_start = df.first_valid_index()
+            last_event_end = df.last_valid_index() + resolution
+            sensor_info = (
+                f" (sensor {sensor.id})" if sensor and hasattr(sensor, "id") else ""
+            )
+
             current_app.logger.warning(
-                f"Prices partially unknown for planning window (sensor {price_sensor.id}). "
+                f"Prices partially unknown for planning window ({sensor_info}). "
                 f"Trimming planning window (from {query_window[0]} until {query_window[-1]}) to {first_event_start} until {last_event_end}."
             )
-            query_window = (first_event_start, last_event_end)
         else:
+            sensor_info = (
+                f" (sensor {df.sensor.id})"
+                if hasattr(df, "sensor") and hasattr(df.sensor, "id")
+                else ""
+            )
             current_app.logger.warning(
-                f"Prices partially unknown for planning window (sensor {price_sensor.id}). "
-                f"Assuming the first price is valid from the start of the planning window ({query_window[0]}), "
-                f"and the last price is valid until the end of the planning window ({query_window[-1]})."
+                f"{capitalize(pluralize(kind_of_values))} partially unknown for planning window ({sensor_info}). "
+                f"Assuming the first {kind_of_values} is valid from the start of the planning window ({query_window[0]}), "
+                f"and the last {kind_of_values} is valid until the end of the planning window ({query_window[-1]})."
             )
             index = initialize_index(
                 start=query_window[0],
                 end=query_window[1],
                 resolution=resolution,
             )
-            price_df = price_df.reindex(index)
-            # or to also forward fill intermediate NaN values, use: price_df = price_df.ffill().bfill()
-            price_df[: price_df.first_valid_index()] = price_df[
-                price_df.index == price_df.first_valid_index()
+            df = df.reindex(index)
+            # or to also forward fill intermediate NaN values, use: df = df.ffill().bfill()
+            df[: df.first_valid_index()] = df[
+                df.index == df.first_valid_index()
             ].values[0]
-            price_df[price_df.last_valid_index() :] = price_df[
-                price_df.index == price_df.last_valid_index()
-            ].values[0]
-    return price_df, query_window
+            df[df.last_valid_index() :] = df[df.index == df.last_valid_index()].values[
+                0
+            ]
+    return df
 
 
 def get_power_values(
@@ -333,6 +368,7 @@ def get_series_from_quantity_or_sensor(
     beliefs_before: datetime | None = None,
     as_instantaneous_events: bool = True,
     boundary_policy: str | None = None,
+    fill_sides: bool = False,
 ) -> pd.Series:
     """
     Get a time series given a quantity or sensor defined on a time window.
@@ -349,6 +385,11 @@ def get_series_from_quantity_or_sensor(
                                     at that time.
     :param as_instantaneous_events: Optionally, convert to instantaneous events, in which case the passed resolution is
                                     interpreted as the desired frequency of the data.
+    :param boundary_policy:         When upsampling to instantaneous events,
+                                    take the 'max', 'min' or 'first' value at event boundaries.
+    :param fill_sides               If True, values are extended to the edges of the query window:
+                                    - The first available value serves as a naive backcast.
+                                    - The last available value serves as a naive forecast.
     :return:                        Pandas Series with the requested time series data.
     """
 
@@ -389,6 +430,11 @@ def get_series_from_quantity_or_sensor(
             f"quantity_or_sensor {variable_quantity} should be a pint Quantity or timely-beliefs Sensor"
         )
 
+    if fill_sides:
+        time_series = extend_to_edges(
+            df=time_series, query_window=query_window, resolution=resolution
+        )
+
     return time_series
 
 
@@ -403,6 +449,7 @@ def get_continuous_series_sensor_or_quantity(
     max_value: float | int | pd.Series = np.nan,
     as_instantaneous_events: bool = False,
     boundary_policy: str | None = None,
+    fill_sides: bool = False,
 ) -> pd.Series:
     """Creates a time series from a sensor, time series specification, or quantity within a specified window,
     falling back to a given `fallback_attribute` and making sure no values exceed `max_value`.
@@ -417,6 +464,9 @@ def get_continuous_series_sensor_or_quantity(
     :param max_value:               Maximum value (also replacing NaN values).
     :param as_instantaneous_events: optionally, convert to instantaneous events, in which case the passed resolution is
                                     interpreted as the desired frequency of the data.
+    :param fill_sides               If True, values are extended to the edges of the query window:
+                                    - The first available value serves as a naive backcast.
+                                    - The last available value serves as a naive forecast.
     :returns:                       time series data with missing values handled based on the chosen method.
     """
     if variable_quantity is None:
@@ -434,6 +484,7 @@ def get_continuous_series_sensor_or_quantity(
         beliefs_before=beliefs_before,
         as_instantaneous_events=as_instantaneous_events,
         boundary_policy=boundary_policy,
+        fill_sides=fill_sides,
     )
 
     # Apply upper limit